add schemas/rpc/block.py
