+++ starknet_py/net/schemas/rpc/state.py
@@ -4,41 +4,22 @@ from marshmallow import EXCLUDE, fields, post_load
 
 from starknet_py.abi.v0.schemas import ContractAbiEntrySchema
 from starknet_py.net.client_models import (
-    BlockHashAndNumber,
-    BlockStateUpdate,
     ContractClass,
-    ContractsNonce,
-    DeclaredContractHash,
     DeployedContract,
+    EmittedEvent,
     EntryPoint,
     EntryPointsByType,
-    PendingBlockStateUpdate,
-    ReplacedClass,
+    Event,
+    EventsChunk,
     SierraContractClass,
     SierraEntryPoint,
     SierraEntryPointsByType,
-    StateDiff,
-    StorageDiffItem,
     SyncStatus,
 )
-from starknet_py.net.schemas.common import (
-    Felt,
-    NonPrefixedHex,
-    NumberAsHex,
-    StorageEntrySchema,
-)
+from starknet_py.net.schemas.common import Felt, NumberAsHex
 from starknet_py.utils.schema import Schema
 
 # pylint: disable=unused-argument, no-self-use
-
-
-class BlockHashAndNumberSchema(Schema):
-    block_hash = Felt(data_key="block_hash", required=True)
-    block_number = fields.Integer(data_key="block_number", required=True)
-
-    @post_load
-    def make_dataclass(self, data, **kwargs) -> BlockHashAndNumber:
-        return BlockHashAndNumber(**data)
 
 
 class SyncStatusSchema(Schema):
@@ -54,19 +35,6 @@ class SyncStatusSchema(Schema):
         return SyncStatus(**data)
 
 
-class StorageDiffSchema(Schema):
-    address = Felt(data_key="address", required=True)
-    storage_entries = fields.List(
-        fields.Nested(StorageEntrySchema()),
-        data_key="storage_entries",
-        required=True,
-    )
-
-    @post_load
-    def make_dataclass(self, data, **kwargs) -> StorageDiffItem:
-        return StorageDiffItem(**data)
-
-
 class ContractDiffSchema(Schema):
     address = Felt(data_key="address", required=True)
     contract_hash = Felt(data_key="contract_hash", required=True)
@@ -74,97 +42,6 @@ class ContractDiffSchema(Schema):
     @post_load
     def make_dataclass(self, data, **kwargs) -> DeployedContract:
         return DeployedContract(**data)
-
-
-class DeclaredContractHashSchema(Schema):
-    class_hash = Felt(data_key="class_hash", required=True)
-    compiled_class_hash = Felt(data_key="compiled_class_hash", required=True)
-
-    @post_load
-    def make_dataclass(self, data, **kwargs) -> DeclaredContractHash:
-        return DeclaredContractHash(**data)
-
-
-class DeployedContractSchema(Schema):
-    address = Felt(data_key="address", required=True)
-    class_hash = NonPrefixedHex(data_key="class_hash", required=True)
-
-    @post_load
-    def make_dataclass(self, data, **kwargs):
-        return DeployedContract(**data)
-
-
-class ContractsNonceSchema(Schema):
-    contract_address = Felt(data_key="contract_address", required=True)
-    nonce = Felt(data_key="nonce", required=True)
-
-    @post_load
-    def make_dataclass(self, data, **kwargs):
-        return ContractsNonce(**data)
-
-
-class ReplacedClassSchema(Schema):
-    contract_address = Felt(data_key="contract_address", required=True)
-    class_hash = Felt(data_key="class_hash", required=True)
-
-    @post_load
-    def make_dataclass(self, data, **kwargs) -> ReplacedClass:
-        return ReplacedClass(**data)
-
-
-class StateDiffSchema(Schema):
-    storage_diffs = fields.List(
-        fields.Nested(StorageDiffSchema()),
-        data_key="storage_diffs",
-        required=True,
-    )
-    deprecated_declared_classes = fields.List(
-        Felt(),
-        data_key="deprecated_declared_classes",
-        required=True,
-    )
-    declared_classes = fields.List(
-        fields.Nested(DeclaredContractHashSchema()),
-        data_key="declared_classes",
-        required=True,
-    )
-    deployed_contracts = fields.List(
-        fields.Nested(DeployedContractSchema()),
-        data_key="deployed_contracts",
-        required=True,
-    )
-    replaced_classes = fields.List(
-        fields.Nested(ReplacedClassSchema()),
-        data_key="replaced_classes",
-        required=True,
-    )
-    nonces = fields.List(
-        fields.Nested(ContractsNonceSchema()), data_key="nonces", required=True
-    )
-
-    @post_load
-    def make_dataclass(self, data, **kwargs) -> StateDiff:
-        return StateDiff(**data)
-
-
-class BlockStateUpdateSchema(Schema):
-    block_hash = Felt(data_key="block_hash", required=True)
-    new_root = Felt(data_key="new_root", required=True)
-    old_root = Felt(data_key="old_root", required=True)
-    state_diff = fields.Nested(StateDiffSchema(), data_key="state_diff", required=True)
-
-    @post_load
-    def make_dataclass(self, data, **kwargs) -> BlockStateUpdate:
-        return BlockStateUpdate(**data)
-
-
-class PendingBlockStateUpdateSchema(Schema):
-    old_root = Felt(data_key="old_root", required=True)
-    state_diff = fields.Nested(StateDiffSchema(), data_key="state_diff", required=True)
-
-    @post_load
-    def make_dataclass(self, data, **kwargs) -> PendingBlockStateUpdate:
-        return PendingBlockStateUpdate(**data)
 
 
 class SierraEntryPointSchema(Schema):
@@ -244,3 +121,36 @@ class ContractClassSchema(Schema):
     @post_load
     def make_dataclass(self, data, **kwargs) -> ContractClass:
         return ContractClass(**data)
+
+
+class EventSchema(Schema):
+    from_address = Felt(data_key="from_address", required=True)
+    keys = fields.List(Felt(), data_key="keys", required=True)
+    data = fields.List(Felt(), data_key="data", required=True)
+
+    @post_load
+    def make_dataclass(self, data, **kwargs) -> Event:
+        return Event(**data)
+
+
+class EmittedEventSchema(EventSchema):
+    transaction_hash = Felt(data_key="transaction_hash", required=True)
+    block_hash = Felt(data_key="block_hash", load_default=None)
+    block_number = fields.Integer(data_key="block_number", load_default=None)
+
+    @post_load
+    def make_dataclass(self, data, **kwargs) -> EmittedEvent:
+        return EmittedEvent(**data)
+
+
+class EventsChunkSchema(Schema):
+    events = fields.List(
+        fields.Nested(EmittedEventSchema()),
+        data_key="events",
+        required=True,
+    )
+    continuation_token = fields.String(data_key="continuation_token", load_default=None)
+
+    @post_load
+    def make_dataclass(self, data, **kwargs):
+        return EventsChunk(**data)
